Support for conditional comments inside build blocks
Making conditional comments (useful for loading css & js targeted at older IE for example) work with usemin build blocks was a pain: if the comments are placed outside the blocks, the build is correct but in development the code is fetched and executed in browser not supporting them; if the comments are inside, they are stripped away when processed.
This commit checks the presence of a conditional comment in each build block with a regExp taken from grunt-usemin code and outputs it around the generated tags.
Some simple tests have been added as well.